fixed the incorrect negative digits in bignum::randomized_div

+++ util.cpp
@@ -429,7 +429,7 @@ bignum bignum::randomized_div(int x) const {
     // strange compiler buug:
     // if I do / and %, function 'divmod' is called, and it complains on launch that divmod is unimplemented
     res.digits[i] = carry / x;
-    carry -= res.digits[i] * x;
+    carry -= res.digits[i] * (long long)(x);
     }
   while(isize(res.digits) && res.digits.back() == 0) res.digits.pop_back();
   if(rand() % x < carry) res += 1;
